backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart
bcrypt==3.2.0